dashboard/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
emoji==2.10.1
apscheduler==3.10.4
jieba==0.42.1
google-generativeai==0.8.3
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
requests==2.32.3
pytest==9.0.2
pytest-cov
httpx==0.25.2
